% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjma_baixar_cjsg.R
\name{tjma_baixar_cjsg}
\alias{tjma_baixar_cjsg}
\title{Baixa consulta jurisprudencial do TJMA}
\usage{
tjma_baixar_cjsg(
  livre = "",
  aspas = FALSE,
  classe = "",
  assunto = "",
  orgao_julgador = "",
  inicio_julg = "",
  fim_julg = "",
  inicio_pb = "",
  fim_pb = "",
  tipo = "A",
  n = NULL,
  diretorio = "."
)
}
\arguments{
\item{livre}{palavra ou texto a ser buscado nas ementas e nos acórdãos}

\item{aspas}{lógico. Colocar a expressão entre aspas?}

\item{classe}{Código da classe processual}

\item{assunto}{Código do assunto}

\item{orgao_julgador}{Código do órgão julgador}

\item{inicio_julg}{data inicial julgamento}

\item{fim_julg}{Data final julgamento}

\item{inicio_pb}{data inicial registro/publicação}

\item{fim_pb}{data final registr/publicacao}

\item{tipo}{"A" Para acórdãos, "D" para decisões monocráticas}

\item{n}{Número de páginas}

\item{diretorio}{Diretório onde serão armazenadas as páginas html}
}
\value{
baixa os htmls das decisões de segunda instância
}
\description{
Baixa consulta jurisprudencial do TJMA
}
\examples{
\dontrun{
baixar_cjsg(livre = "Lei Maria da Penha")
}

}
\keyword{tjsp,acórdãos}
