% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjma_baixar_cjsg.R
\name{tjma_baixar_cjsg}
\alias{tjma_baixar_cjsg}
\title{Baixar jurisprudência do TJMA}
\usage{
tjma_baixar_cjsg(
  chave = "",
  data_inicio = "",
  data_fim = "",
  sistema = "0",
  relator = "0",
  revisor = "0",
  camara = "0",
  tipo_pesquisa = "1",
  diretorio = "."
)
}
\arguments{
\item{chave}{Campo busca livre. Se não informado, irá buscar
palavra "tribunal"}

\item{data_inicio}{Data inicial da publicação no formato dd/mm/aaaa}

\item{data_fim}{Data final da publicação no formato dd/mm/aaaa}

\item{sistema}{ThemisSG ou PGE2G}

\item{relator}{Nome do relator}

\item{revisor}{Nome do revisor}

\item{camara}{Câmara}

\item{tipo_pesquisa}{1 para ementa, 2 para advogado,
3 para número do acórdão, 4 para número do processo}

\item{diretorio}{Diretório onde armazenar os dados}
}
\value{
As tabelas são convertidas em rds e salvas no
diretório informado
}
\description{
Baixar jurisprudência do TJMA
}
